Move the toolkit's internal tests out of the source Otherwise it messes with the bundler and lets `vitest` seep into the non-Vitest bundle
